Remove unused functionality
 - Renamed Tone.c to beep.c.
 - Beeps can no longer be played forever.
 - Removed noTone function to stop current beep.

+++ keyboard/y63b/matrix.c
@@ -27,7 +27,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "util.h"
 #include "matrix.h"
 
-extern void beep(unsigned int frequency, unsigned long duration);
+extern void beep(uint32_t frequency, uint32_t duration);
 
 
 #ifndef DEBOUNCE
